schemas: Update display_defaults.branch_label for Auspice v2 config and dataset
Reconciles the two schemas with some minor improvements + a bug fix.
While the enum values in one schema were valid, they're not the _only_
valid values.

This display_default refers to the keys of branch_attr.labels objects on
nodes, so I reused the former's pattern of allowed characters.  I think
that pattern may be too restrictive and not accurately describe what
Auspice will handle, but that's a broader issue with other parts of the
schema too and something to investigate for another day.

Hint that there's a special-cased value ("none") by mentioning it
explicitly in the pattern.