Error when any group-by column is not found
Actualizing the TODO removed in this change:

> TODO: We should consider treating this case as an actual error and exiting here with a nonzero code.

With 2342f07b75f2ba9953cab8254ebef3d5febb5b5d, raising FilterException(AugurError) alone is sufficient to exit with a nonzero code.

Also updated error text to remove message "No sequences-per-group sampling will be done", which was part of a warning. Implicitly, nothing will be done when exiting with an ERROR.