fix: allows nuc annotation to be pulled in through `.gff`
Currently, `augur translate` cannot output `nuc` annotation, if `.gff` is used as input. This is due to the features being loaded being overly restrictive to genes.

Allowing `source` to be read in as features allows the `.gff` to contain nuc annotations in the following format:
```
MT903344.1	Genbank	source	1	197233	.	+	.	locus_tag=nuc
```

That way `augur export doesn't crash if a `.gff` is used as annotation input for translate.