[ancestral] Send warnings to stderr
Input trees with non-labelled internal nodes always printed a WARNING
but this was going to stdout rather than stderr. This in turn allowed
the tests to pass when they shouldn't have.

Note: I think this warning should be a fatal error in almost every
conceivable case. To describe ancestral sequences without knowing which
node they correspond to in the input tree is a recipe for confusion.