Fix missing groups for included strains
Fixes a subtle bug when grouping with subsampled max sequences and
force-included strains that could potentially be considered for
subsampling. In the first pass through the metadata, we remove
force-included strains from consideration for subsampling. However, we
do not similarly remove these strains in the second pass through the
metadata. If these strains have not been filtered earlier, we consider
them for subsampling in the second pass and look for a queue with their
group-by attributes. Since queue creation depends on the group
attributes we find in the first pass, it is possible for the group
attributes of the force-included strains to not have a queue.

This commit adds a functional test to recreate the bug (originally
discovered in the ncov workflow) and updates the logic in the second
pass to remove force-included strains prior to building subsampling
groups.