Allow commands to add their own subparser
Instead of just having each command to add their own arguments, allow
them add their own subparser. This gives each command complete control
of its own parser and their subparsers to customize them as needed.
This allows each command to define its own command name instead
of parsing the command name from the module name, which is important for
the following commit to rename the `import` module.

This commit also introduces the behavior to print a command's help
when called if it does not have its own `run` function.

Note that for a majority of the commands, `register_arguments` was
updated to `register_parser`. However, for `align`, `filter` and `mask`,
I kept the `register_arguments` functions since they are used in
unit testing.

This change came from discussion of how to extract the boilerplate for
registering subparsers so that commands can use the same function to
register their own subcommands.¹

¹ https://github.com/nextstrain/augur/pull/1002#discussion_r917222854