Refactor measurements module to a package
Moves the measurements module to its own package where the subcommands
are split out into their own modules. The subcommand subparsers are
added with the new `add_command_subparsers` function.

This is an effort to reorganize augur commands with subcommands as
packages instead of continuously adding subcommands as modules.
Starting with the measurements command since it is a new command that is
unlikely to already be used by anyone outside of the Nextstrain. In the
future, we may want to consider refactoring export and validate
in the same way. Putting that off for now since reorganizing those
modules may result in breaking outside uses of their APIs.