Refactor import_ and import_beast to a package
Moves the import_ and import_beast modules to an import_ package and
uses the new `add_command_subparsers` function to add the beast
subparser.

Search via cs.github.com¹ shows that there are no external uses of
the `augur.import` or `augur.import_beast` APIs, so this should be
safe to move to a package.

¹ https://cs.github.com/?scopeName=All+repos&scope=&q=%22augur.import%22