refactor: rename `import.py` to `import_.py`
Since `import` is a Python keyword, the best practice is to name the
module as `import_` according to PEP-8 guidelines.¹

This should not affect the usage of the `augur import` command since
the command name remains unchanged. Since the module only has
`register_parser` and `run` functions, there is unlikely to be any
external uses of `augur.import`.

This change came from suggestion by @tsibley in review.²

¹ https://peps.python.org/pep-0008/#descriptive-naming-styles
² https://github.com/nextstrain/augur/pull/1002#discussion_r918328749