Add functional tests for translate
Adds Cram tests for augur translate with GenBank- and GFF-based gene
maps. Much of the data are copied from the Zika and TB builds. These
tests capture the current behavior of augur translate in preparation for
future improvements to this behavior.

This commit adds support to `diff_jsons.py` for excluding paths in JSON
files by regular expression, so we can ignore differences in temporary
file names. Version DeepDiff 5.2.0 introduced a command line interface
[1] that provides this same functionality and more, eliminating the need
for custom scripts like `diff_jsons.py`. In the future, we may wish to
update Augur's deepdiff dependency [2] and switch to this CLI.

[1] https://zepworks.com/deepdiff/current/commandline.html
[2] https://github.com/nextstrain/augur/blob/67027a6ae6a7a324c8eb646593cc5c0a89650ccf/setup.py#L72