Write out the actual sequence id for translations
Uses the actual sequence id for the reference that corresponds to the
coordinates written by augur translate, instead of writing out the
"seqid" as the filename of the gene map. The file name does not always
represent the id of the sequence (e.g., when the name is "genemap.gff"),
whereas the sequence id in the given GenBank or GFF record should be
more accurate.