Report actual missing filename for index paths
Fixes a confusing error message produced by augur index when the output
directory for the sequence index does not exist but the error message
says the sequences file is missing. The fix is to use the actual
filename associated with the `FileNotFoundError` to report the "missing"
file, instead of assuming that the missing file is the input.

Adds a functional test to capture this expected behavior.