curate: Add shared parser and metadata input option
The shared parser will be used to house the shared options for
all subcommands. I have not found a "non-hacky" way to add parent
parsers outside of the `add_parser` call, so each subcommand must add
the shared parser to their own parser within their `register_parser`
function.

This includes the first shared option for curate commands, which is
the metadata input file. The functional test for metadata input run
the `passthru` subcommand, but it should cover all future subcommands
because the inputs for all subcommands are processed the same way.