io/metadata: Add `read_table_to_dict`
Reads a CSV or TSV file and yields each row as a dictionary.
Uses the csv standard library instead of pandas to avoid having to
convert the pandas DataFrame to dicts. This is written as a general
I/O function, but the intention is to use it to parse metadata inputs
for the new `augur curate` command so that records can be easily
streamed and modified by each subcommand.