types: Add enum `DataErrorMethod`
Add an enum of string values that represent how a data error should be
handled. The values are currently just limited to "error_first",
"error_all", "warn" and "silent", which are the common patterns of how
we handle data errors in Augur.

This was added with the intention to be used to dictate how the new
`augur curate` command should handle duplicate records in the following
commit.