curate: Add combined metadata/fasta input options
It is not clear yet if we want to add deduplication as an `augur curate`
command, but for now we can report duplicate records in both metadata
and FASTA files to the users. Commands will error on the first duplicate
or unmatched record by default since we usually expect unique records in
the metadata and FASTA files that are 1:1.

Note that if there are duplicate sequence ids in the FASTA file, pyfastx
will always return the first sequence in the file.

Note that the command can still produce an output even when an error is
raised due to unmatched and duplicate records. However, users should be
able to see the error message if running the command manually and
workflow managers such as Snakemake will delete the output if the
command exits with an error.