io/sequences: Add `write_record_to_fasta`
Writes sequences to a FASTA file by parsing the sequence id and sequence
from an iterator that yields dict records. After writing the sequence
to the FASTA file, the function yields a copy of the record without the
sequence field so that the metadata can be consumed downstream.

This is written as a general I/O function, but the intention is to use
it to write the FASTA output for the new `augur curate` command before
the records are output to a metadata file.