[clades] export labels as specific branch labels
Previously clade membership (i.e. the coloring) and the branch labels
defining the root of the clade were defined via:
<OUTPUT_NODE_DATA> → nodes → <node_name> → clade_membership, and
<OUTPUT_NODE_DATA> → nodes → <node_name> → clade_annotation.
`augur export` would then convert the clade_annotation into a
branch label named 'clade'.

Here we change the format of augur clade's OUTPUT_NODE_DATA so
that the membership and labels are now stored via:
<OUTPUT_NODE_DATA> → nodes → <node_name> → clade_membership, and
<OUTPUT_NODE_DATA> → branches → <node_name> → labels → clade.
The previous commit modified augur export to handle this format.

Augur pipelines should be fully backwards compatible as long as a new
major version of augur is released, as we ensure that node-data files
are created by the same augur (major) version. Scripts which relied on
the format of this node-data file may be affected.

Note that we keep the key 'clade_membership' deliberately: this is
used in auspice-config JSONs and auspice URLs, and so changing it will
cause lots of downstream issues for a minimal syntax improvement. (The
`clade_annotation` key name was never exported in auspice JSONs.)

This commit paves the way for allowing custom key names.