io/metadata: Allow IO Buffer as input for read_metadata_to_dict
Previously, `read_metadata_to_dict` only allowed files as inputs because
it needed to reset the file position after taking a sample for the
csv.Sniffer.

This commit allows IO buffer inputs by converting the sample string back
to a text stream that is then chained with the original text stream.