Fix compatibility with Pandas 1.5.0
Pandas 1.5.0 was released earlier this morning and moved an exception
class we use.  Routine CI caught this backwards-incompatible change
affecting our error handling for `augur filter --query …` evaluation.

I audited for usage of other exceptions moved in the same upstream
change (as noted in the release notes) and found none in Augur.

Resolves: <https://github.com/nextstrain/augur/issues/1049>
Related-to: <https://github.com/pandas-dev/pandas/issues/27656>