io/metadata: Add duplicate reporting to read_table_to_dict
We do not need to handle de-duplication in this input function, but
we can at least report duplicate records if requested. If no id column
is provided, it will use the first column of the input file to identify
duplicates.

The duplicate reporting methods include:
* error_first = error on the first duplicate record encountered
* error_all = run through all records and error with a message that
includes all duplicate ids
* warn = emits a warning on every encountered duplicate and prints a
final warning message that is a summary of all duplicate ids
* silent = ignore duplicates