curate: Add metadata duplicate reporting options
It is not clear yet if we want to add deduplication as an `augur curate`
command, but for now we can at least report duplicate records to the
users. Commands will error on the first duplicate by default since we
usually expect metadata to not contain duplicate records.

Note that when "error_all" is requested, the command will still produce
an output. This is because the error will only be raised after the
generator has been exhausted. Since `curate` is chain of generators, the
input generator will only be exhausted by writing the output. However,
users should be able to see the error message if running the command
manually and workflow managers such as Snakemake will delete the output
if the command exits with an error.