Use stacked ambiguous date checking
Previously, the ambiguous date checks were 1:1 with the generated date
columns. This should not be the case since month needs to check for
ambiguous year, and week needs to check for anything that is ambiguous.

Separate the ambiguous date checking from the column generation, and
update the conditions for the former to be "stacking" (i.e. year is
always checked, month is checked for month/week, and day is checked for
week only).