io/metadata: Add duplicates reporting to `read_metadata_with_sequences`
The function now errors on records with duplicate sequence ids since we
expect the provided metadata and FASTA files to have unique sequence ids.
This can be toggled to warnings or silenced.

The unmatched and duplicates reporting methods are separate parameters
to allow users to fine tune which errors get reported. The reporting is
structured to make sure the any warnings are emitted before the error
is raised. If both unmatched and duplicate reporting request ERROR_ALL,
then they are raised as a single error to give the user feedback on both.