Fix trait inference when tips have missing values
Fixes a bug in trait inference when any tips are missing data (i.e.,
have values of "?" in the field for which discrete traits are being
reconstructed) that caused the last known trait value to be treated as
missing data, too. This commit adds a functional test that confirms the
bug and then resolves the bug by excluding the missing data value from
the traits passed to `reconstruct_discrete_traits`. That function then
properly handles reconstruction of just the missing data values when the
missing data value is not part of the traits provided.

Note that the original issue stemmed from how the TreeTime wrapper
function `reconstruct_discrete_traits` behaved when missing data were
included in the traits it received [1]. This fix was inspired by
the way TreeTime's `mugration` function [2] avoided the bug in TreeTime
itself.

This commit also adds a functional test to confirm that when all values
in a discrete trait column are missing values ("?") the traits command
issues a warning that no discrete traits were found instead of saying
that a single trait of "?" was found.

[1] https://github.com/neherlab/treetime/pull/216
[2] https://github.com/neherlab/treetime/blob/a27cc685a89f9e0cb3b2d9608d82de3f1e48ae1d/treetime/wrappers.py#L919-L920