Add functional test for DRM file column validation
Note 1: This is the first functional test of sequence-traits, so there
are some unnecessary warnings captured in the output. Keep them for now,
and they can be addressed by future improvements.

Note 2: I'm trying a new way of writing Cram tests by not changing
directory to stay in the per-test temporary directory, and referencing
any existing files relative to $TESTDIR.

Note 3: The other arguments to sequence-traits are mostly arbitrary and
may not be logically correct, however the intent is just to test
column validation, which works.