docs: Add usage docs for augur curate
Add usage docs for the new augur curate subcommand as a subdirectory
to separate the docs for each sub-subcommand because we are expecting
to add more in the future. The I/O options are repeated for each
sub-subcommand because they use a shared parser. I think of this as
feature to show how each sub-subcommand can be a single command.

Once we add more sub-subcommands, we can add examples of curate
piplines to the main curate docs page. We can also add examples of
curate interacting with external tools such as tsv-utils or csvtk.