Try using titers with values below threshold
Attempts to include more titer data into models by detecting measurement
values that are below a given threshold (e.g., "<40") and converting
those values to a reasonable numerical value that is one two-fold
dilution value less than the threshold (e.g., 20, in the example above).
This change allows us to at least include these values in our titer
models and the measurements panel, so we have a complete picture of all
the data available. If this general approach makes sense, we might
consider a similar approach for values _above_ a given threshold, too.