Serialize common numpy data types in write_json
There is a possibility that data to be written by this function contains
numpy data types that are un-serializable by the default JSONEncoder¹.
Make them serializable using a custom subclass.

Note that defining a default function for un-serializable values is
another approach², but using a custom subclass is more complete since it
allows for proper fall-back to the default encoder which automatically
raises meaningful TypeErrors.

¹ https://docs.python.org/3/library/json.html#json.JSONEncoder
² https://stackoverflow.com/a/65151218/4410590