Fix pivot logic to always include the end date
Fixes the previous, flawed pivot logic that allowed pivots to _not_
include the end date by a substantial margin (for example: given an end
date of January 6, 2023 and a 3-month interval, pivots ended at November
1, 2022). The new implementation uses a third-party date library,
dateutil, that supports month-based time deltas in addition to the
week-based deltas supported by most other Python date libraries. This
new implementation ensures that the end date always appears as the last
entry in the the pivot outputs (as we always wanted it to), works
backward through time to create pivots from that end date minus the
given time delta interval, and stops when the next pivot occurs before
the start date. With respect to the start date, this behavior is similar
to the prior implementation which never guaranteed that the start date
would be included as a pivot unless the pivot interval allowed it.

A key technical detail of this new implementation is that the internal
and intermediate pivot datetime instances are converted to floating
point values with the `numeric_date` function instead of the
`timestamp_to_float` function. The first of these calculates decimal
values from the fraction of days in a year, while the second calculates
decimal values from the fraction of months in a year. The day-based
fractions from `numeric_date` are already used throughout Augur and
TreeTime, while `timestamp_to_float` has only been used for frequency
and distance calculations. Since other functions like
`float_to_datestring` also interpret decimals as fractions of days in a
year, the change to pivot logic here makes the output of `get_pivots`
more consistent with the rest of Augur's date ecosystem.