setup-miniconda: Use Mambaforge
conda-libmamba-solver without any channels is fast, but when used with
python-version, it installs Python from the defaults channel which can
be unreliable compared to conda-forge.

An equally fast alternative is Mambaforge, which is a Miniforge variant
that comes with Mamba in the base environment. It uses the Mamba within
the setup-miniconda step for faster package resolution speeds when used
with python-version and the conda-forge channel.