export v2: Add --validation-mode={error,warn,skip} option
Extends the existing --skip-validation flag to be more nuanced by the
addition of a "warn" mode.  Our validation is improving, but not always
correct.  The "warn" mode allows a workflow to emit messages about
validation failures, but keep going in the face of them instead of
aborting the whole workflow due to a non-zero exit code.

We may want to change the default mode to "warn"—at least until our
validation failures have few "false positives" or "low-consequence
positives" where the output files still load ok in Auspice—but I'll
leave this for separate consideration.

The --skip-validation flag is now an alias for --validation-mode=skip.

Related-to: <https://github.com/nextstrain/augur/issues/1044>