Add InvalidDate exception class to handle user errors
This makes it easier to handle user errors with meaningful messages.

The exception class can't be in dates.py and referenced in
date_disambiguator.py since that would result in a circular import. A
better pattern would be to refactor dates.py to a package, but I'll
leave that for another time.