refine: Remove --year-bounds
This is the only usage of the `min_max_year` parameter in several
internal functions where it is passed around ultimately to
DateDisambiguator. Upon inspection, the value is set on the object but
is actually unused. This is because some old logic got lost in a
refactor: 1cf41fb7eb388077aff0ded044d32aab2e4c5590

Since this option is not doing anything and seems to be unused for the
most part¹, remove it as well as all the supporting code.

¹ https://github.com/search?q=augur+refine+--year-bounds+NOT+path%3A%2F%5Eaugur%5C%2Frefine.py%2F&type=code