Remove intermediate function ambiguous_date_to_date_range()
It was only a one-line alias to another function call. Removed it to
reduce extra complexity in code.

Updated tests to use the caller function. It returns a numeric date, so
make a 3-way comparison between the returned value, the datetime.date,
and the exact float representation of the numeric date. The only
difference here is that the first test no longer goes through
DateDisambiguator so it doesn't return a range.