curate formate-dates: Mask failed date strings
By default, completely mask date strings with "XXXX-XX-XX" for dates
that failed date formatting so that they are still in the proper
ISO 8601 dates format for downstream Augur commands. Users can turn off
the masking with the added `--no-mask-failure` option.

Note the `store_false` actions produce semi-misleading docs in the help
output, so the new option suppresses the default value in the help message
with `SKIP_AUTO_DEFAULT_IN_HELP` as suggested by @tsibley in review.