measurements export: Allow multiple threshold values
Updates the option `--threshold` to accept multiple values.
If a single "threshold" value is provided via the `--collection-config`,
it will be converted to a "thresholds" array in the measurements JSON.

Note, that this will produce measurements JSONs that will only have the
"thresholds" array but not the deprecated single "threshold" value. This
means the thresholds will not be displayed in Auspice v2.42.0 or lower
because Auspice still expects the single "threshold" value.