measurements export: Update option to `--thresholds`
The default argparse abbreviation matching¹ allows us to update the
option to `--thresholds` to match the key in the produced JSON while
still supporting the previous single `--threshold` option.

Updated one of the Cram tests to use the plural form while leaving the
others in their original singular form to ensure both work as expected.

¹ https://docs.python.org/3.7/library/argparse.html#argument-abbreviations-prefix-matching