Calculate all pivot points based on end pivot
Previously, monthly pivots calculated from the end of a month failed to
account for the variation in the number of days per month. This was due
to subtle differences between subtracting an isodate.Duration against a
decremental date (previous implementation) and subtracting an
isodate.Duration with a multiplier against a fixed date (new
implementation).

Tests updated with the new behavior.