docs: Update usage docs for `ancestral` and `translate`
Adds a high level explanation based on an email sent by @huddlej for
explaining the entropy panel in Auspice. (Visible in the Nextstrain
discussion forum¹)

Includes a generic example of the output node-data JSONs so users can
produce their own JSONs outside of Augur if desired.

The general summary of how the commands work are added to the module
docstrings so that they can be available in the command line help output.
The example JSONs seem like detail that would clutter this help output
so they are added to the rst docs.

¹ https://discussion.nextstrain.org/t/make-auspice-entropy-panel-show-mutations-with-respect-to-arbitrary-sequence-reference-rather-than-reconstructed-root-of-the-tree/1333/2