filter: Add `--empty-output-reporting={error,warn,silent}` option
Allows users to specify reporting behavior if filtering produces an
empty result. The default behavior is still to raise an error if
there's an empty result.

This change was prompted by the refactoring of the Nextstrain
seasonal influenza builds, where we use `augur filter` to filter
sequences for a specific region.¹ The lack of B Victoria sequences in
the Japan/Korea region in the past 2 years leads to an empty filtered
result that raises an error, which stops the entire pipeline.²
The downstream rules are able to handle an empty filtered result, so
we just need an option to silence the error.

¹ https://github.com/nextstrain/seasonal-flu/blob/de7875364a52d05cfa6c66cc49d53a5f8b2de1a0/profiles/private.nextflu.org/export.smk#L67-L87
² https://bedfordlab.slack.com/archives/C03KWDET9/p1677875939132259?thread_ts=1677792435.138489&cid=C03KWDET9