Add test for export with color by metadata
Adds a functional test and mock data that should work where augur export
includes the `--color-by-metadata` argument with a column named
`custom_trait`. The resulting Auspice JSON should include the
`custom_trait` as a node attribute for each tip and as a coloring that
appears after the colorings explicitly defined in the Auspice config
JSON.

The original goal of this test was to confirm that "color by metadata"
colorings disrupt the order of all colorings in the Auspice config JSON,
but this test also reveals a separate potential bug where one of the
node data annotations (`mutation_length`) that also appears in the
metadata gets dropped completety from the Auspice JSON when the color by
metadata argument is used.