devel/test: remove backticked example command
The script failed during the release process on 2023-03-13 and @tsibley
noticed the backticked example command is not escaped so the script
will actually install Mamba leading to a "Conda-solver-OOM"¹

This commit removes the example command to avoid this issue. We will
revisit the "proper" solution to consolidate the CI and release jobs
in the future.

¹ https://bedfordlab.slack.com/archives/C01LCTT7JNN/p1678745783771769?thread_ts=1678305212.677099&cid=C01LCTT7JNN