devel/test: Install mamba if not available
Previously assumed the Conda solver was crashing because of the
unintended Mamba installation (see b103a3824d0a5d20c8ef951b31902e0becf84fe5).
However, the script still crashed in the release job after removing
the example command.¹

This commit updates the script to install Mamba if it's not available
and then create the environment with mamba to get around the Conda
solver issue.

¹ https://github.com/nextstrain/augur/actions/runs/4410444018/jobs/7727917621