[clades] allow node-data nodes to be a subset of tree nodes
Our current implementation of read_node_data requires that every
node in the tree is specified in the (merged) node_data files. For
mutations this is overkill -- many nodes don't have mutations and it's
overkill to require node_data JSONs to specify things like
`"node_name": {"muts": []}`.

This may well be the general behaviour we want, but i didn't want to
modify the read_node_data function which sees extensive use.

A welcome side effect of these changes is that we no longer have to
supply both nuc and aa_muts.