Add DEFAULT_AUGUR_RECURSION_LIMIT=10000
We've had to use the `AUGUR_RECURSION_LIMIT` environment variable to
bump the recursion limit for SARS-CoV-2 builds¹, mpox builds², and
our `nextstrain-jobs` definition in AWS Batch. From a general search of
the envvar across GitHub³, 10,000 looks like the default that others use
as well (although this may be influenced by our discussions⁴).

¹ https://github.com/nextstrain/ncov/pull/690
² https://github.com/nextstrain/monkeypox/pull/150
³ https://cs.github.com/?scopeName=All+repos&scope=&q=AUGUR_RECURSION_LIMIT%3D
⁴ https://github.com/nextstrain/augur/issues/328