Do not produce duplicate coloring entries
The logic for creating colorings involves many steps, and when adding a
coloring we had neglected to check if it had already been added.
For instance, colorings explicitly defined in the auspice-config
which were provided via node-data files (rather than the metadata tsv)
would be added twice. In our current nCoV build there are 7 such
duplicates! There were also a few duplicates that were present in our
tests. Whoops! These weren't a problem for auspice (the first was used),
but added unnecessary complexity and bytes to the JSON.

Closes #719