Restore clades to top entry in colorings
The intention of the coloring logic is that if an auspice-config provides
the clade_membership key then it is exported at that position in the
colorings list. If clade_membership is not explicitly set in the config
(but is present in a node-data file) then we have (for a very long time)
added it as the very first entry in the colorings list.

PR #728 (augur v22.0.0) erroneously modified the behavior of the second
case described above, which has now been restored by this commit.