curate: simplify uses of `DataErrorMethod` enum
Now that `DataErrorMethod` is an `ArgparseEnum`, we no longer have to
specify `value` for the argparse choices and we do not have to convert
the argument strings back to enum members.

Specify type as `DataErrorMethod.argtype` in order to get better
error messages when the provided option value is not a valid choice.