Use mypy to check types
The codebase has sprinkles of typing here and there. Previously, they
were not checked in any way and purely for extra context to readers of
the code. Now, they are checked and enforced through mypy.

There are several mypy errors when run against the code as-is. The
following commits will address those errors.

Setup was largely copied from existing mypy usage in Nextstrain CLI,
including the way mypy is run via pytest.